morse/requirements.txt
pycw
pydub
python-zbase32
